Finalizing

+++ src/main/java/org/example/beadando_gy/Entity/Eredmeny.java
@@ -12,20 +12,20 @@ public class Eredmeny {
     private Long id;
 
     private LocalDate datum;
-    private Long pilotaaz;
+    private Integer pilotaaz; // megfelel az int(11) a DB-ben
     private Integer helyezes;
     private String hiba;
     private String csapat;
     private String tipus;
     private String motor;
 
-    // getterek/setterek
+    // getter/setter
     public Long getId() { return id; }
     public void setId(Long id) { this.id = id; }
     public LocalDate getDatum() { return datum; }
     public void setDatum(LocalDate datum) { this.datum = datum; }
-    public Long getPilotaaz() { return pilotaaz; }
-    public void setPilotaaz(Long pilotaaz) { this.pilotaaz = pilotaaz; }
+    public Integer getPilotaaz() { return pilotaaz; }
+    public void setPilotaaz(Integer pilotaaz) { this.pilotaaz = pilotaaz; }
     public Integer getHelyezes() { return helyezes; }
     public void setHelyezes(Integer helyezes) { this.helyezes = helyezes; }
     public String getHiba() { return hiba; }
